requirements.txt
affine==2.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.1
cycler==0.12.1
docopt==0.6.2
fonttools==4.53.1
geopandas==1.0.1
ghp-import==2.1.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
mergedeep==1.3.4
mkdocs==1.6.0
mkdocs-get-deps==0.2.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pydap==3.4.1
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
rasterio==1.3.10
requests==2.32.3
rioxarray==0.17.0
ruamel.yaml==0.16.2
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.1
scipy==1.14.0
setuptools==69.5.1
shapely==2.0.5
six==1.16.0
snuggs==1.4.7
soupsieve==2.5
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
WebOb==1.8.7
wheel==0.43.0
xarray==2024.7.0
us==3.2.0
earthengine-api==1.0.0
SALib==1.5.1
jupyter==1.0.0
notebook==7.0.3
redis
shortuuid
mkdocs>=1.5
mkdocs-material
mkdocs-bootstrap4